requirements.txt
matplotlib==3.3.1
git+https://github.com/LIAAD/yake
numpy==1.18.5
pandas==1.0.5
spacy==2.3.2
wordcloud==1.8.0
streamlit==0.64.0
Pillow==7.2.0
python-bidi==0.4.2
